backend/userauths/serializers.py
from rest_framework import serializers
from userauths.models import User, Profile


# serializers converts pyhona object or any complex data into Json

class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = User
        fields = "__all__"
        # exclude = 'email'


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = "__all__"

    def to_representation(self, instance):
        response = super().to_representation(instance)
        response['user'] = UserSerializer(instance.user).data
        return response